fix(build): don't copy the bower_components folder in build
This changes the default copy of the `bower_components` to not copy it
during `grunt build` to the */dist* folder. **bowerInstall** should copy
the needed files into the *index.html* so **usemin** can process them.
If the files don't confirm to the Bower spec, a user can manually add a
line in the *index.html* file so they are included in the build process.

Fixes #590

BREAKING CHANGE: bower_components is no longer copied into the /dist
folder.