Update readme.md
I just built with this and used the "yo angular:route myRoute --uri=my/route" from the readme. I found the new view at the uri "/myRoute" so I updated the doc readme to reflect it, "/#/myRoute" might be even more helpful for some so they can tell that it needs the hash.