fix(build): do not clean git related subfolders of dist
clean scope was updated in 9db87bf6f61252db604e9ae4e9b13360f8b96eef, but
the scope of excluded git related files was not updated accordingly.
This caused clean to keep only the .git folder, but none of its contents.
I am not sure if this was caused by a change in node-glob, but it does
not work as expected.

I discovered this issue when using grunt-build-control, which will check
for the existence of a .git subfolder in dist and assume a repo exists
if such a folder exists (that check could probably be made more robust).