fix(app) copy dist fonts and ensure revs update
Fonts were not being copied for builds since the fonts dir was
moved to /styles/fonts. Updated the path in the copy:dist task.

Closes #963

Fonts were not being copied for builds since the fonts dir was moved to `/styles/fonts`. Updated the path in the `copy:dist` task.

Not 100% sure how the `usemin` task works, but without `/styles` in the `assetsDirs`, it wasn't updating the references to the reved fonts.