chore(jshint): drop deprecated jshint options
According to JSHint documentation they are limiting their scope to
issues of code correctness, and are deprecating some style rules.

The list of options that are deprecated but used in our .jshintrc
are

* camelcase
* immed
* newcap
* quotemark

This patch drops these options from the .jshintrc templates and also
the actual .jshintrc used by this project itself.

JSHint points to the JSCS — JavaScript Code Style project for future
style checking.

See http://jshint.com/docs/options for a list of all options of
JSHint including deprecation warnings.