chore(release): simplify and replace release
Remove the stage task since the config was removed and it was only used
for staging the changelog.

Removed grunt-release as the only thing it was needed for was npm
publishing and GH releases. conventional-github-releaser takes care of
the GH part and a custom grunt task takes care of the npm part.

Replace the custom bumping and semver with grunt-bump